Lesson2end

+++ Lesson2/lesson_2_task_3.py
@@ -5,5 +5,5 @@ def square(side):
     return math.ceil(side * side)
 
 
-square_side = int(input("Введите сторону кадрата: "))
+square_side = float(input("Введите сторону кадрата: "))
 print(f"Площадь квадрата: {square(square_side)}")
